requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Tau-Phah-Ji-Command==1.0.0
anthropic>=0.40.0
python-dotenv==1.0.0
gunicorn==21.2.0
pypinyin==0.55.0
supabase>=2.0.0
